try to isolate issue with compile flags being moved by the wrapper scripts
When calling thee gcc_warpper shell script, the script parses through
the arguments for CROSS_CFLAGS, CROSS_CXXFLAGS, and moves them acording
to some internal rules. This causes an agument looking like:

-mllvm -shuffle-stack-frames

to become

-mllvm SOME_ARBITRARY ARGUMENTS -c -shuffle-stack-frames

This leads to a build failure, and because this argument unwrapping is
hidden in the script the real commpiler command is not displayed, so
that it appears as though the compiler does not understand its
directive.

In reality this ordered set of arguments has been split such that they
will always fail.

TODO: determine where our compiler arguments should go.

current solution points them directly after compiler invocation within
the wrapper scripts.

Other options are to modify the wrapper script to correctly process
these arguments.