Add support for code layout randomization
Run-time code layout randomization is a very involved defense
that includes compile time instrumentation via LLVM, link-time
information with the help of Selfrando's TrapLinker, and a run-time
component through compiler-rt. ENABLE_CODE_LAYOUT_RANDO can be used
in the config to enable this defense. Note that this requires the
Multicompiler.